+++ CHANGELOG.md
@@ -3,6 +3,7 @@
  - Use scriptable object for audio unit
  - Add more documentation comments
  - Make pool holder as don't destroy on load
+ - Add custom editor and property drawer for core scripts
 
 # 0.1.0
  - Import base logics of this repository